backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
PyPDF2
python-docx
google-generativeai
pydantic
pydantic-settings
python-dotenv
aiofiles
httpx 